fix: update docker compose to run as a nonroot user
There is a new tests which uses permissions to validate temp directory handing of files and this won't work as root. On top of this it is good practice to use a non root user so adjusting this will be a net positive.

This requires changing from mounting the sources to adding them to the container prior to execution of tests, which may add a bit more time to the build.